+++ hal/__init__.py
@@ -1,0 +1,12 @@
+from hal.arm_interface import ArmInterface, JointStates, Pose
+from hal.camera_interface import CameraInterface, CameraIntrinsics
+from hal.gripper_interface import GripperInterface
+
+__all__ = [
+    "ArmInterface",
+    "JointStates",
+    "Pose",
+    "GripperInterface",
+    "CameraInterface",
+    "CameraIntrinsics",
+]